requirements.txt
django
djangorestframework
environs
psycopg2-binary
dj-database-url
whitenoise[brotli]
gunicorn
uvicorn